chore: restructure command handling and add LSP commands
