+++ .golangci.yml
@@ -1,0 +1,54 @@
+version: "2"
+run:
+  allow-parallel-runners: true
+linters:
+  default: none
+  enable:
+    - copyloopvar
+    - dupl
+    - errcheck
+    - ginkgolinter
+    - goconst
+    - gocyclo
+    - govet
+    - ineffassign
+    - lll
+    - misspell
+    - nakedret
+    - prealloc
+    - revive
+    - staticcheck
+    - unconvert
+    - unparam
+    - unused
+  settings:
+    revive:
+      rules:
+        - name: comment-spacings
+  # exclusions:
+  #   generated: lax
+  #   rules:
+  #     - linters:
+  #         - lll
+  #       path: api/*
+  #     - linters:
+  #         - dupl
+  #         - lll
+  #       path: internal/*
+  #   paths:
+  #     - third_party$
+  #     - builtin$
+  #     - examples$
+formatters:
+  enable:
+    - gci
+    - gofmt
+    - gofumpt
+    - goimports
+    - golines
+  # exclusions:
+  #   generated: lax
+  #   paths:
+  #     - third_party$
+  #     - builtin$
+  #     - examples$
