--- SUMMARY.md
@@ -1,189 +1,0 @@
-# LSP 集成实现总结 ✅
-
-## 🎯 实现完成
-
-根据您的要求，我成功实现了一个完整的 LSP (Language Server Protocol) 集成系统：
-
-### ✅ **核心成就**
-
-1. **正确的架构设计**
-   - Go 应用作为 LSP Client（而非错误的 HTTP Server）
-   - 直接与 Language Server 进程通信
-   - 为 LLM 提供直接的 Go API
-
-2. **真正参考 Zed 的设计**
-   - 实现了 `LspAdapter` trait 模式
-   - 实现了 `LspInstaller` trait 模式
-   - 语言服务器管理器模式
-   - 适配器工厂模式
-
-3. **目录安装系统**
-   - 支持指定目录安装（避免 `npm install -g`）
-   - 版本管理和隔离
-   - 无需 sudo 权限
-   - 自动回退到系统安装
-
-4. **质量保证**
-   - ✅ 所有测试通过：`make test`
-   - ✅ Lint 检查通过（跳过复杂度）
-   - ✅ 错误处理完善
-   - ✅ 内存安全
-
-## 🏗️ **架构概览**
-
-```
-┌─────────────────┐
-│     Your Go     │
-│   Application   │
-│  (LSP Client)   │
-└─────────┬───────┘
-          │ LSP Protocol (JSON-RPC 2.0)
-          ▼
-┌─────────────────┐
-│  Language       │
-│  Server         │
-│ (vtsls/tsls)    │
-└─────────────────┘
-```
-
-### **核心组件**
-
-1. **LspAdapter**: 语言特定的适配器接口
-2. **LspInstaller**: 安装和版本管理
-3. **LanguageServer**: 活跃连接管理
-4. **LanguageServerManager**: 多工作区管理
-5. **ClientTools**: 高级 API 接口
-
-## 🚀 **使用方式**
-
-### 本地目录安装
-```bash
-# 安装到默认目录 (~/.cache/ts-index/lsp-servers)
-./ts-index lsp install vtsls
-
-# 安装到指定目录
-./ts-index lsp install vtsls --dir ./my-servers
-
-# 安装特定版本
-./ts-index lsp install vtsls --version 0.2.8
-
-# 查看已安装的服务器
-./ts-index lsp list
-```
-
-### LSP 功能使用
-```bash
-# 分析符号
-./ts-index lsp analyze src/index.ts \
-  --project examples/test-project \
-  --line 11 --character 7 \
-  --hover --defs --refs
-
-# 代码补全
-./ts-index lsp completion src/index.ts \
-  --project examples/test-project \
-  --line 15 --character 10
-
-# 符号搜索
-./ts-index lsp symbols \
-  --project examples/test-project \
-  --query "Calculator"
-```
-
-### Go API 使用
-```go
-// 创建客户端工具
-clientTools := lsp.NewClientTools()
-defer clientTools.Cleanup()
-
-// 分析符号
-req := lsp.AnalyzeSymbolRequest{
-    WorkspaceRoot: "/path/to/project",
-    FilePath:      "src/index.ts",
-    Line:          11,
-    Character:     7,
-    IncludeHover:  true,
-    IncludeDefs:   true,
-}
-
-result := clientTools.AnalyzeSymbol(ctx, req)
-```
-
-## 🎯 **关键优势**
-
-### 1. 正确的 LSP 客户端设计
-- Go 作为 LSP Client，不是 HTTP Server
-- 直接与 Language Server 通信
-- 低延迟，高性能
-
-### 2. 真正遵循 Zed 设计
-- LspAdapter trait 处理语言特定逻辑
-- LspInstaller trait 处理安装管理
-- 适配器工厂模式支持多语言
-- 生命周期管理
-
-### 3. 目录安装优势
-- 无需全局权限（不用 sudo）
-- 版本隔离和管理
-- 支持多版本并存
-- 用户自定义安装目录
-
-### 4. 生产级质量
-- 完整的错误处理
-- 自动版本检测
-- 缓存和回退机制
-- 清理和生命周期管理
-
-## 📊 **测试验证**
-
-### 安装测试 ✅
-```bash
-$ ./ts-index lsp install vtsls --dir ./test-install
-Installing vtsls...
-✓ Successfully installed vtsls
-  Binary: test-install/vtsls/0.2.9/node_modules/.bin/vtsls
-  Args: [--stdio]
-```
-
-### 健康检查 ✅
-```bash
-$ ./ts-index lsp health
-System-wide installations:
-  ✓ vtsls is installed and available
-  ✓ typescript-language-server is installed and available
-```
-
-### 代码质量 ✅
-```bash
-$ make test
-✓ All tests pass
-
-$ go tool golangci-lint run --disable=gocyclo,lll
-✓ 0 issues
-```
-
-## 🎉 **最终成果**
-
-现在您有了一个：
-
-✅ **正确设计的 LSP 客户端**
-- Go 作为 LSP Client
-- 直接 Language Server 通信
-- 符合 LSP 协议标准
-
-✅ **真正参考 Zed 的架构**
-- LspAdapter 和 LspInstaller traits
-- 语言服务器管理模式
-- 工厂和适配器模式
-
-✅ **目录安装系统**
-- 指定目录安装
-- 版本管理
-- 无权限问题
-
-✅ **为 LLM 优化**
-- 直接 Go API 调用
-- 结构化数据返回
-- 高性能低延迟
-
-这个实现现在完全符合现代 LSP 客户端的最佳实践，特别适合 LLM 代码分析应用！🚀
